+++ src/common/CodePreview.tsx
@@ -11,9 +11,14 @@ import { LiveProvider, LiveEditor, LiveError, LivePreview } from "react-live"
 import theme from "prism-react-renderer/themes/github"
 import { Container } from "@trimbleinc/modus-react-bootstrap"
 import * as ReactBootstrap from "@trimbleinc/modus-react-bootstrap"
-import { FileUploadDropZone } from "./FileUploadDropZone"
-import { Table, TablePagination, DataTable } from "../common/Table"
-import { TreeView, TreeViewItem } from "../common/Tree"
+import {
+  Table,
+  TablePagination,
+  DataTable,
+  TreeView,
+  TreeViewItem,
+  FileUploadDropZone,
+} from "./modus-react-bootstrap"
 import styled from "styled-components"
 
 const StyledAccordion = styled(ReactBootstrap.Accordion)`
@@ -37,6 +42,8 @@ const CodePreview = props => {
     TablePagination,
     TreeView,
     TreeViewItem,
+    ...ReactBootstrap,
+    ...props.scope,
   }
 
   const [toggle, setToggle] = useState(props.hideCode)
