--- src/common/FileUploadDropZone/index.tsx
@@ -1,2 +1,0 @@
-export { default as FileUploadDropZone } from "./FileUploadDropZone"
-export type { FileUploadDropZoneProps } from "./FileUploadDropZone"
