--- src/common/Tree/types.tsx
@@ -1,5 +1,0 @@
-import { TreeItem } from "."
-
-export type TreeItemExtended = TreeItem & {
-  element?: any
-}
