+++ src/common/modus-react-bootstrap/TablePaginationStyled.tsx
