+++ src/common/modus-react-bootstrap/index.tsx
@@ -4,6 +4,8 @@ import {
   UseSortByColumnOptions,
 } from "react-table"
 
+import { ReactNode } from "react"
+
 export { default as Table } from "./Table"
 export type { TableProps } from "./Table"
 
@@ -20,3 +22,21 @@ export type { DataTableProps } from "./DataTable"
 
 export { default as TablePagination } from "./TablePagination"
 export type { TablePaginationProps } from "./TablePagination"
+
+export { default as FileUploadDropZone } from "./FileUploadDropZone"
+export type { FileUploadDropZoneProps } from "./FileUploadDropZone"
+
+export { default as TreeView } from "./TreeView"
+export type { TreeViewProps } from "./TreeView"
+
+export { default as TreeViewItem } from "./TreeViewItem"
+export type { TreeViewItemProps } from "./TreeViewItem"
+
+export type TreeItem = {
+  id: number
+  parentId: number
+  index: number
+  label?: ReactNode
+  children?: TreeItem[]
+  disabled?: boolean
+}
