+++ src/pages/sample/content-tree.tsx
@@ -1,7 +1,6 @@
 import * as React from "react"
 import { Container, Row } from "@trimbleinc/modus-react-bootstrap"
-import TreeViewItem from "../../common/Tree/TreeViewItem"
-import TreeView from "../../common/Tree/TreeView"
+import { TreeView, TreeViewItem } from "../../common/modus-react-bootstrap"
 import { ModusIconsScripts } from "../../common/ExternalDependencyHelper"
 import "../../assets/css/main.scss"
 
@@ -14,7 +13,7 @@ const ContentTreePage = props => {
             <div className="col-12 col-lg-6 pt-5 mt-xl-5">
               <div style={{ width: "400px" }}>
                 <ModusIconsScripts />
-                <TreeView multiSelectCheckBox>
+                <TreeView multiSelectCheckBox id="test1">
                   <TreeViewItem nodeId={1} label="Layout">
                     <TreeViewItem nodeId={2} label="Main Layout">
                       <TreeViewItem nodeId={3} label="Header"></TreeViewItem>
